Fix IoAlign requirement, add volumeInfo list for future use
I want to keep a list of volumes and such

+++ Common/Kernel/Disk/Efi/Efi.h
@@ -15,7 +15,7 @@
   [[nodiscard]] bool InitializeDiskSubsystem_Efi(void);
   bool TerminateDiskSubsystem_Efi(void);
 
-  uint64 GetBlockSize_Efi(uint64 DrivePosition);
-  [[nodiscard]] bool ReadDisk_Efi(void* Pointer, uint64 Lba, uint64 NumBlocks, uint64 DrivePosition);
+  uint64 GetBlockSize_Efi(uint16 DrivePosition);
+  [[nodiscard]] bool ReadDisk_Efi(void* Pointer, uint64 Lba, uint64 NumBlocks, uint16 DrivePosition);
 
 #endif
